When adding a new comment for a post, thge user is redirected to the post page and the comment is saved as draft. It will be displaied when it will be approved.
Changes for this are in:

_comments - Display only comments that are approved.
comment/_form - added a hidden field for storing the post id and call the comment/create
Comment model - beforeSave method was modified. Now the default status of a Comment is 1. Also, the new Datetime was replaced with date()
CommentController - Ading in the comment post_id the id of the post from the $_POST request. Also, redirect this to the post page.
post/view - added a new parameter (post) to the /comment/_form partial view